third_party/WebKit/LayoutTests/external/wpt/css/css-shapes/parsing/shape-outside-invalid-position-expected.txt

This is a testharness.js-based test.
FAIL e.style['shape-outside'] = "circle(at center left 1px)" should not set the property value assert_equals: expected "" but got "circle(at 1px 50%)"
FAIL e.style['shape-outside'] = "circle(at center top 2px)" should not set the property value assert_equals: expected "" but got "circle(at 50% 2px)"
FAIL e.style['shape-outside'] = "circle(at right 3% center)" should not set the property value assert_equals: expected "" but got "circle(at 97% 50%)"
FAIL e.style['shape-outside'] = "circle(at left 4px top)" should not set the property value assert_equals: expected "" but got "circle(at 4px 0%)"
FAIL e.style['shape-outside'] = "circle(at right top 5px)" should not set the property value assert_equals: expected "" but got "circle(at 100% 5px)"
FAIL e.style['shape-outside'] = "circle(at bottom 6% center)" should not set the property value assert_equals: expected "" but got "circle(at 50% 94%)"
FAIL e.style['shape-outside'] = "circle(at bottom 7% left)" should not set the property value assert_equals: expected "" but got "circle(at 0% 93%)"
FAIL e.style['shape-outside'] = "circle(at bottom right 8%)" should not set the property value assert_equals: expected "" but got "circle(at 92% 100%)"
Harness: the test ran to completion.
